Get setup on windows
Need to uninstall anaconda and when re-installing it, add it to the path so git bash can recognize it. Then install git bash when prompted to do so after attempting to open the repo in a CLI. Then might have to run `conda init` and or `conda update conda` from within git bash to complete the config. Then from GitHub Desktop software can open the repo in Git Bash and issue conda commands and run the repo (after some minor windows path adjustements to the source code :-)